Change how filtering and scrolling work
We now filter only when you mouseover the MR reference. This stops the
activity timeline going crazy as you mouse around the MRs.

We also pin the filter when you click the MR reference. This means that
you can click to filter, then scroll around the timeline.

This removes the need to have the timeline always shown, so we remove
the scrolling on the MR cards.